broke up login function into login-function file
