+++ auth/login-controller.js
@@ -1,289 +1,11 @@
-const bcrypt = require("bcryptjs");
-const Users = require("../users/user-model.js");
-const Genres = require("../author-content/genres-model");
-const Libraries = require("../content-library/library-model");
-
-const { genToken } = require("./generate-token");
+const { LoginFuncDisplay, loginFuncEmail } = require("./login-function");
 
 exports.postLogin = [
   (req, res) => {
-    const { email, password } = req.body;
     if (req.body.login.indexOf("@") !== -1) {
-      Users.findByEmail(req.body.login)
-        .first()
-        .then((u) => {
-          if (u && bcrypt.compareSync(password, u.password)) {
-            const token = genToken(u);
-            const userList = {
-              id: u.id,
-              userType: u.user_type,
-              firstName: u.first_name,
-              lastName: u.last_name,
-              displayName: u.display_name,
-              email: u.email,
-              country: u.country,
-              state: u.state,
-              city: u.city,
-              image: u.image,
-              createdAt: u.created_at,
-            };
-            Libraries.findByIdLibrary(u.id).then((library) => {
-              const libraryContent = library.map((ele) => {
-                const genres = [];
-                const objectArray = Object.entries(ele);
-                objectArray.map(([key, value]) => {
-                  if (value === true) {
-                    return genres.push(key);
-                  }
-                });
-                const {
-                  id,
-                  user_id,
-                  title,
-                  description,
-                  img_url,
-                  content_url,
-                  created_at,
-                  last_updated,
-                  public_id,
-                } = ele;
-                const newObj = {
-                  id,
-                  user_id,
-                  title,
-                  description,
-                  img_url,
-                  content_url,
-                  created_at,
-                  last_updated,
-                  public_id,
-                  genres,
-                };
-                return newObj;
-              });
-              Users.findAgentInfoId(u.id).then((agentInfo) => {
-                const agentInfoList = agentInfo.map((info) => {
-                  const {
-                    agent_type,
-                    agency_type,
-                    agency_address,
-                    agency_phone_number,
-                    agency_email,
-                  } = info;
-                  return {
-                    agent_type,
-                    agency_type,
-                    agency_address,
-                    agency_phone_number,
-                    agency_email,
-                  };
-                });
-                Genres.findById(u.id)
-                  .then((content) => {
-                    const contentGenre = content.map((ele) => {
-                      const genres = [];
-                      const objectArray = Object.entries(ele);
-                      objectArray.map(([key, value]) => {
-                        if (value === true) {
-                          return genres.push(key);
-                        }
-                      });
-                      const {
-                        id,
-                        user_id,
-                        title,
-                        description,
-                        img_url,
-                        content_url,
-                        created_at,
-                        last_updated,
-                        public_id,
-                        img_public_id,
-                      } = ele;
-                      const newObj = {
-                        id,
-                        user_id,
-                        title,
-                        description,
-                        img_url,
-                        content_url,
-                        created_at,
-                        last_updated,
-                        public_id,
-                        img_public_id,
-                        genres,
-                      };
-                      return newObj;
-                    });
-                    if (u.user_type === "author") {
-                      res.status(200).json({
-                        User: userList,
-                        AuthorContent: contentGenre,
-                        ContentLibrary: libraryContent,
-                        Token: token,
-                      });
-                    } else if (u.user_type === "agent") {
-                      res.status(200).json({
-                        User: userList,
-                        AgentInfo: agentInfoList,
-                        contentLibrary: libraryContent,
-                        Token: token,
-                      });
-                    } else {
-                      res.status(200).json({
-                        User: userList,
-                        contentLibrary: libraryContent,
-                        Token: token,
-                      });
-                    }
-                  })
-                  .catch((err) => {
-                    res.status(500).json(err);
-                  });
-              });
-            });
-          } else {
-            res.status(400).json({ message: "Invalid Credentials" });
-          }
-        });
+      loginFuncEmail(req, res);
     } else {
-      Users.findByDisplayName(req.body.login)
-        .first()
-        .then((u) => {
-          if (u && bcrypt.compareSync(password, u.password)) {
-            const token = genToken(u);
-            const userList = {
-              id: u.id,
-              userType: u.user_type,
-              firstName: u.first_name,
-              lastName: u.last_name,
-              displayName: u.display_name,
-              email: u.email,
-              country: u.country,
-              state: u.state,
-              city: u.city,
-              image: u.image,
-              createdAt: u.created_at,
-            };
-            Libraries.findByIdLibrary(u.id).then((library) => {
-              const libraryContent = library.map((ele) => {
-                const genres = [];
-                const objectArray = Object.entries(ele);
-                objectArray.map(([key, value]) => {
-                  if (value === true) {
-                    return genres.push(key);
-                  }
-                });
-                const {
-                  id,
-                  user_id,
-                  title,
-                  description,
-                  img_url,
-                  content_url,
-                  created_at,
-                  last_updated,
-                  public_id,
-                  img_public_id,
-                } = ele;
-                const newObj = {
-                  id,
-                  user_id,
-                  title,
-                  description,
-                  img_url,
-                  content_url,
-                  created_at,
-                  last_updated,
-                  public_id,
-                  img_public_id,
-                  genres,
-                };
-                return newObj;
-              });
-              Users.findAgentInfoId(u.id).then((agentInfo) => {
-                const agentInfoList = agentInfo.map((info) => {
-                  const {
-                    agent_type,
-                    agency_type,
-                    agency_address,
-                    agency_phone_number,
-                    agency_email,
-                  } = info;
-                  return {
-                    agent_type,
-                    agency_type,
-                    agency_address,
-                    agency_phone_number,
-                    agency_email,
-                  };
-                });
-                Genres.findById(u.id)
-                  .then((content) => {
-                    const contentGenre = content.map((ele) => {
-                      const genres = [];
-                      const objectArray = Object.entries(ele);
-                      objectArray.map(([key, value]) => {
-                        if (value === true) {
-                          return genres.push(key);
-                        }
-                      });
-                      const {
-                        id,
-                        user_id,
-                        title,
-                        description,
-                        img_url,
-                        content_url,
-                        created_at,
-                        last_updated,
-                        public_id,
-                      } = ele;
-                      const newObj = {
-                        id,
-                        user_id,
-                        title,
-                        description,
-                        img_url,
-                        content_url,
-                        created_at,
-                        last_updated,
-                        public_id,
-                        genres,
-                      };
-                      return newObj;
-                    });
-                    if (u.user_type === "author") {
-                      res.status(200).json({
-                        User: userList,
-                        AuthorContent: contentGenre,
-                        ContentLibrary: libraryContent,
-                        Token: token,
-                      });
-                    } else if (u.user_type === "agent") {
-                      res.status(200).json({
-                        User: userList,
-                        AgentInfo: agentInfoList,
-                        contentLibrary: libraryContent,
-                        Token: token,
-                      });
-                    } else {
-                      res.status(200).json({
-                        User: userList,
-                        contentLibrary: libraryContent,
-                        Token: token,
-                      });
-                    }
-                  })
-                  .catch((err) => {
-                    res.status(500).json(err);
-                  });
-              });
-            });
-          } else {
-            res.status(400).json({ message: "Invalid Credentials" });
-          }
-        });
+      LoginFuncDisplay(req, res);
     }
   },
 ];
